Tabs to spaces, atomic operations for dictionary entry removal.

+++ Megaboggle/Timer.cpp
@@ -3,15 +3,15 @@
 
 Timer::Timer()
 {
-	mStartTime = std::chrono::duration_cast<std::chrono::milliseconds>(
-		std::chrono::system_clock::now().time_since_epoch()
-		).count();
+    mStartTime = std::chrono::duration_cast<std::chrono::milliseconds>(
+        std::chrono::system_clock::now().time_since_epoch()
+        ).count();
 }
 
-long Timer::stop()
+long long Timer::stop()
 {
-	long stopTime = std::chrono::duration_cast<std::chrono::milliseconds>(
-		std::chrono::system_clock::now().time_since_epoch()).count();
+    long long stopTime = std::chrono::duration_cast<std::chrono::milliseconds>(
+        std::chrono::system_clock::now().time_since_epoch()).count();
 
-	return stopTime - mStartTime;
+    return stopTime - mStartTime;
 }
